fix: reflect changes
reflect TokenRepository, TokenProvider

+++ src/main/java/kr/kro/calcking/calckingwebbe/services/auth/SignOutService.java
@@ -11,32 +11,34 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import kr.kro.calcking.calckingwebbe.providers.CookieProvider;
-import kr.kro.calcking.calckingwebbe.providers.JWTProvider;
-import kr.kro.calcking.calckingwebbe.repositories.UserTokenRepository;
+import kr.kro.calcking.calckingwebbe.repositories.TokenRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class SignOutService {
-  private final UserTokenRepository userTokenRepository;
+  private final TokenRepository tokenRepository;
   private final CookieProvider cookieProvider;
-  private final JWTProvider jwtProvider;
 
   // POST (/sign-out)
-  public ResponseEntity<Map<String, Object>> signOut(
-      HttpServletRequest request, HttpServletResponse response) {
-    Map<String, Object> responseMap = new HashMap<>();
-
-    // 로그아웃 로직
+  public ResponseEntity<Map<String, Object>> signOut(HttpServletRequest request, HttpServletResponse response) {
+    // 로그아웃
     Cookie refreshCookie = cookieProvider.getCookie("refresh_token", request);
-    userTokenRepository
-        .deleteUserTokenByUID(
-            jwtProvider.getUIDFromAccessToken(request.getHeader("Authorization").substring(7)));
-    cookieProvider.deleteCookie(refreshCookie, response);
+    String refreshToken = cookieProvider.getTokenFromCookie(refreshCookie);
+    if (refreshCookie != null) {
+      cookieProvider.deleteCookie(refreshCookie, response);
+      tokenRepository.deleteToken(refreshToken);
+    }
 
-    // JSON 응답 로직
-    responseMap.put("message", "로그아웃 성공!");
-    responseMap.put("status", String.valueOf(HttpStatus.OK));
-    return ResponseEntity.status(HttpStatus.OK).body(responseMap);
+    // 성공 응답
+    return responseMap("로그아웃 성공!", HttpStatus.OK);
+  }
+
+  // JSON 응답 생성 은닉 메서드
+  private ResponseEntity<Map<String, Object>> responseMap(String message, HttpStatus status) {
+    Map<String, Object> responseMap = new HashMap<>();
+    responseMap.put("message", message);
+    responseMap.put("status", String.valueOf(status));
+    return ResponseEntity.status(status).body(responseMap);
   }
 }
